This now links
+ removed samle c code out of includes cause the main function
  clashed with the ghc generated one
+ remove package.yaml file automatically (so you don't accidently edit it)
+ deelte awk stuff, no longer necisary because we figured out c2hs
