+++ app/exe.hs
@@ -6,3 +6,6 @@ import           Cut.SphinxBindings
 main :: IO ()
 main = error "compiles mostly"
   -- detect_words "heyo.raw"
+
+fun :: ResultCode
+fun = SUCCESS
